Respect PathBase for Swagger-UI.
The SwaggerMiddleware already adds the Request.PathBase property to its path. The SwaggerUi-project did not yet do this. This resulted in a broken UI because the path to swagger could not be resolved properly.

I've added a simple middleware to the Basic-website to test this. This middleware allows to make a request with a "X-Forwarded-PathBase", which gets set as the PathBase.

We have this exact setup with our APIs. They are hosted behind a reverse-proxy which sends the path under which they are hosted as a HTTP header.